Enrich procedure details the same way as functions
