+++ pkg/sdk/procedures_ext.go
@@ -29,9 +29,15 @@ type ProcedureDetails struct {
 	Handler                    *string // present for python, java, and scala (hidden when SECURE)
 	RuntimeVersion             *string // present for python, java, and scala (hidden when SECURE)
 	Packages                   *string // list // present for python, java, and scala (hidden when SECURE)
-	TargetPath                 *string // list present for scala and java (hidden when SECURE)
+	TargetPath                 *string // present for scala and java (hidden when SECURE)
 	InstalledPackages          *string // list present for python (hidden when SECURE)
 	ExecuteAs                  string  // present for all procedure types
+
+	NormalizedImports    []NormalizedPath
+	NormalizedTargetPath *NormalizedPath
+	ReturnDataType       datatypes.DataType
+	ReturnNotNull        bool
+	NormalizedArguments  []NormalizedArgument
 }
 
 func procedureDetailsFromRows(rows []ProcedureDetail) (*ProcedureDetails, error) {
@@ -71,6 +77,37 @@ func procedureDetailsFromRows(rows []ProcedureDetail) (*ProcedureDetails, error)
 			v.TargetPath = row.Value
 		}
 	}
+	if e := errors.Join(errs...); e != nil {
+		return nil, e
+	}
+
+	if normalizedImports, err := parseFunctionOrProcedureImports(v.Imports); err != nil {
+		errs = append(errs, err)
+	} else {
+		v.NormalizedImports = normalizedImports
+	}
+
+	if v.TargetPath != nil {
+		if p, err := parseFunctionOrProcedureStageLocationPath(*v.TargetPath); err != nil {
+			errs = append(errs, err)
+		} else {
+			v.NormalizedTargetPath = p
+		}
+	}
+
+	if dt, returnNotNull, err := parseFunctionOrProcedureReturns(v.Returns); err != nil {
+		errs = append(errs, err)
+	} else {
+		v.ReturnDataType = dt
+		v.ReturnNotNull = returnNotNull
+	}
+
+	if args, err := parseFunctionOrProcedureSignature(v.Signature); err != nil {
+		errs = append(errs, err)
+	} else {
+		v.NormalizedArguments = args
+	}
+
 	return v, errors.Join(errs...)
 }
